requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Deprecated==1.2.7
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pathspec==0.7.0
PyGithub==1.45
PyJWT==1.7.1
python-dotenv==0.10.5
regex==2020.1.8
requests==2.22.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
